Remove dummy files from app
This commit will fail in CI, because there are no actual tests to be run,
thus, no coverage can be collected, resulting in coverage below the defined
threshold of 95%, resulting in a failing build.

Coverage will be pushed as soon as possibel. Hint: When there is actual code
to be tested.