Remove mypy type checking
Typing of the Django-related codebase is a nightmare, up until now without any
benefit.

As of now, mypy is removed from the pre-commit hooks, as there is no point in
ignoring all Django-stuff (by comment in file or excluding the file in the
pre-commit configuration).

Keep typing in mind when there is more, non-Django code.