Hide the unavailable EventProvider
The PluginWidget was modified to use a HiddenInput for the unavailable
event providers.

This is the most simple solution to hide this from the user, as he is not
expected to interact with this field anyway.
There may be another implementation in the future. The user *should* be
informed about the unavailability of a formerly selected plugin, probably
even with the possibility to clean those plugins from the user's profile.
Good enough for now!