Fix CalenderEntry.category
The CalenderEntry instance must provide not only the (translated) label,
but the actual value aswell.
This enables filtering within the templates against known strings.

Additionally: I could make this change without breaking unittests. This is
really bad, as this is one of the core components of the app. I guess this
was not revealed earlier, because the line was marked with "pragma: nocover",
which is bad aswell.