Add layout specific (config) form template
As of Django 4.0, forms are rendered using templates, which allows finer
control of *how* a form should be presented to the user.

With this commit, the new technique is applied to the layout configuration
process.
The current implementation is backwards compatible to Django 3.1/3.2.

See #40 for more details. #40 is intended to apply this technique to all
forms of django-calingen.