Squash migrations before release
There were two migrations. They are not actually squashed using Django's
admin command "squashmigrations", but the initial migration is recreated
by deleting the existing ones.

This was done, because this will be the actual first release to PyPI, so
existing projects are not an issue. For future reference: Use
"squashmigrations"!